Сделал автосоздание таблицы в случае удаления бд
