+++ LMS/account.html
@@ -11,6 +11,8 @@
     <link rel="stylesheet" type="text/css" href="sources/style/account.css">
     <link rel="stylesheet" type="text/css" href="sources/style/message.css">
     <link rel="stylesheet" type="text/css" href="sources/style/button.css">
+
+    <script type="text/javascript" src="sources/js/nesql.js"></script>
     <script type="text/javascript" src="sources/js/add.js"></script>
     <script type="text/javascript" src="sources/js/message.js"></script>
 </head>
@@ -40,7 +42,6 @@
     </div>
 
     <script>
-        let fs = require("fs"); // Подключаем библиотеку для работы с файловой системой
         let GET = parseURL(); // Сюда сохраняем данные из ссылки на страницу
 
         let users = new Table('users'); // Создаем таблицу для работы с учениками
