+++ LMS/add_book.html
@@ -11,6 +11,8 @@
     <link rel="stylesheet" type="text/css" href="sources/style/button.css">
     <link rel="stylesheet" type="text/css" href="sources/style/message.css">
     <link rel="stylesheet" type="text/css" href="sources/style/add.css">
+
+    <script type="text/javascript" src="sources/js/nesql.js"></script>
     <script type="text/javascript" src="sources/js/add.js"></script>
     <script type="text/javascript" src="sources/js/add_bk.js"></script>
     <script type="text/javascript" src="sources/js/add_book.js"></script>
