+++ LMS/index.html
@@ -5,6 +5,7 @@
 
 <head>
     <meta charset="UTF-8">
+    <meta http-equiv="Content-Security-Policy" content="script-src 'self';" />
     <title>Главная</title>
 
     <link rel="stylesheet" type="text/css" href="sources/style/index.css">
