+++ LMS/search_student.html
@@ -10,6 +10,8 @@
     <link rel="stylesheet" type="text/css" href="sources/style/button.css">
     <link rel="stylesheet" type="text/css" href="sources/style/search.css">
     <link rel="stylesheet" type="text/css" href="sources/style/message.css">
+
+    <script src="sources/js/nesql.js"></script>
     <script src="sources/js/search.js"></script>
     <script src="sources/js/message.js"></script>
 </head>
