+++ LMS/sources/js/add_bk.js
@@ -8,74 +8,6 @@ window.onload = function() {
     let bookID  = document.querySelector("input[name=id]");
     let comment = document.querySelector("textarea[name=comment]");
 
-    let fs = require("fs");
-
-    // Мой код, принимающий массив с данными и преобразующий его в блоки
-    class Table {
-        constructor(name) {
-            this.name = name + '.json';
-        }
-
-        SELECT() {
-            try {
-                return JSON.parse(fs.readFileSync(this.name, 'utf8'));
-            } catch (err) {
-                console.log('ОШИБКА ЧТЕНИЯ ТАБЛИЦЫ', err);
-            }
-        }
-
-        INSERT(values) {
-            let params = this.SELECT();
-
-            for (let i = 0; i < Object.keys(params).length-1; i++) {
-                if(values[i] === null) values[i] = null;
-            }
-
-            let len = params['id']['users'].length;
-            let i = -1;
-            let nextID = len ? params['id']['users'][len-1]+1 : 1;
-
-            for (let item in params) {
-                if (item == 'id') {
-                    params[item]['users'][len] = nextID;
-                } else {
-                    params[item]['users'][len] = values[i];
-                }
-                i++;
-            }
-
-            this.write(params);
-            return nextID;
-        }
-
-        get nextGroupID() {
-            let MaxGroupID = 0;
-            for (let item of this.SELECT()['groupid']['users']) {
-                if (item && item > MaxGroupID) {
-                    MaxGroupID = item;
-                }
-            }
-            return MaxGroupID+1;
-        }
-
-        isSameID(otherItem) {
-            for (let item of this.SELECT()['inventoryno']['users']) {
-                if (otherItem == item) {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        write(params) {
-            try {
-                fs.writeFileSync(this.name, JSON.stringify(params));
-            } catch (err) {
-                console.log('ОШИБКА ИЗМЕНЕНИЯ ТАБЛИЦЫ', err);
-            }
-        }
-    }
-
     let books = new Table('books');
 
     document.querySelector('#submit').onclick = function() {
